VishwamAI-main/requirements.txt
dm-haiku
jax
transformers
tensorflow
torch
onnx
onnx2keras
tf-keras
Pillow
numpy
scipy
requests
PyMuPDF
datasets
pydub
speechrecognition
